requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
chromadb
openai
python-dotenv
numpy
requests
